requirements.txt
# Flask Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Caching==2.1.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Async Tasks
celery==5.3.2
redis==5.0.0

# HTTP Requests
requests==2.31.0
urllib3==2.0.4

# Discord
discord.py==2.3.2

# Data Processing
pandas==2.1.1
numpy==1.25.2

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.13.0

# Environment
python-dotenv==1.0.0

# Validation
marshmallow==3.20.1

# Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0

# Development
flake8==6.0.0
black==23.7.0

# Production
gunicorn==21.2.0
